Issue #2: Move all the wrappers to the 'core' folder.
