+++ src/pages/AppRoot/AppRoot.tsx
@@ -1,5 +1,5 @@
-import { AppWrapper } from 'src/ui/wrappers/AppWrapper';
-import { AppRouterWrapper } from 'src/ui/wrappers/AppRouterWrapper';
+import { AppWrapper } from 'src/core/wrappers/AppWrapper';
+import { AppRouterWrapper } from 'src/core/wrappers/AppRouterWrapper';
 
 export function AppRoot() {
   return (
